fix(examples): Update scalar_add_1 to new API
This commit updates the `scalar_add_1` example to align with the recent changes to the PJRT C++ API. The `execute` and `toHost` methods have been updated to
handle the new data structures.

Additionally, a Python script for generating the StableHLO has been modified to write directly to a `.stablehlo` file, and the associated CMake and C++
files have been updated to use the new filename.